Extract `FeedImageDataLoaderSpy` into a shared scope to remove duplication.
